Create components for search bar and new search query button
